// sdk/node/src/generate/margin/order/model_cancel_all_orders_by_symbol_resp.ts
// Code generated by Kucoin Universal SDK Generator; DO NOT EDIT.

import { instanceToPlain, Exclude, plainToInstance } from 'class-transformer';
import { RestResponse } from '@model/common';
import { Response } from '@internal/interfaces/response';

export class CancelAllOrdersBySymbolResp
    implements Response<CancelAllOrdersBySymbolResp, RestResponse>
{
    /**
     *
     */
    data?: string;
    /**
     * common response
     */
    @Exclude()
    commonResponse?: RestResponse;

    setCommonResponse(response: RestResponse): void {
        this.commonResponse = response;
    }

    fromJson(input: string): CancelAllOrdersBySymbolResp {
        const jsonObject = JSON.parse(input);
        return plainToInstance(CancelAllOrdersBySymbolResp, { data: jsonObject });
    }

    toJson(): string {
        return JSON.stringify(instanceToPlain(this.data));
    }

    fromObject(jsonObject: Object): CancelAllOrdersBySymbolResp {
        return plainToInstance(CancelAllOrdersBySymbolResp, { data: jsonObject });
    }
}